Update use of destructuring object and fixed export default problem
