+++ src/index.js
@@ -1,8 +1,7 @@
 /* eslint-disable */
 import { renderAddTodo } from './add-todo';
 import { projectDisplayListener } from './functions';
-import { showProjects } from './displayProjects';
-import { showTodos, showTodosInClickedProject } from './displayTodos';
+import { showTodos, showTodosInClickedProject, showProjects } from './displayTodos';
 import { Storage } from './storage';
 
 const todoApp = (() => {
